Use `extend/os` pattern.
