+++ Library/Homebrew/extend/os/search.rb
@@ -1,0 +1,1 @@
+require "extend/os/mac/search" if OS.mac?
